Adds `DaggerAndroidClassTransform` to transform for dagger android
- Adds `DaggerTrackClassTransform` interface for handling transformation
