+++ dagger-track/src/main/java/me/amanjeet/daggertrack/transform/DaggerTrackClassTransform.kt
@@ -1,0 +1,8 @@
+package me.amanjeet.daggertrack.transform
+
+import javassist.CtClass
+import java.io.File
+
+internal interface DaggerTrackClassTransform {
+    fun handleClassTransformation(allCtClasses: List<CtClass>, outputDir: File)
+}
